Slight formatting and keybind prep

+++ screeny.py
@@ -13,6 +13,7 @@ import time
 import webbrowser
 from base64 import b64encode
 import pyperclip
+from keybind import KeyBinder
 
 from tkinter import filedialog
 from tkinter import *
@@ -33,18 +34,23 @@ icon = open(appdata + "\Screeny\icon.ico")
 icons = itertools.cycle(glob.glob('*.ico'))
 hover_text = "SCREENY - Screenshot Assistant\n\nUse Win+Shift+S to use default mode\nRight click for all modes"
 
-def hello(sysTrayIcon):
+
+# ----------- GENERAL -------------
+
+
+def blank(sysTrayIcon):
     pass
     
 def close(sysTrayIcon): 
-    os.remove("image.png")
+    if os.path.exists("image.png"):
+        os.remove("image.png")
 
 def clearClip():
     if windll.user32.OpenClipboard(None):
         windll.user32.EmptyClipboard()
         windll.user32.CloseClipboard()
 
-def runSS(sysTrayIcon):
+def getSS(sysTrayIcon):
     keyboard.press(Key.cmd)
     keyboard.press(Key.shift_l)
     keyboard.press('s')
@@ -52,18 +58,25 @@ def runSS(sysTrayIcon):
     keyboard.release(Key.shift_l)
     keyboard.release(Key.cmd)
 
-
-def ocrSave(sysTrayIcon):
+def grabSS():
     clearClip()
     time.sleep(0.2)
     image = PIL.ImageGrab.grabclipboard()
-    runSS(None)
+    getSS(None)
     print("Loading: ", end="")
     while image is None:
         print(".", end="")
         image = PIL.ImageGrab.grabclipboard()
         time.sleep(0.1)
     print("\nDone")
+    return image
+
+# ------------- TOOLS -------------
+
+# -------- OCR TOOLS
+
+def ocrSave(sysTrayIcon):
+    image = grabSS()
     print("\nText:")
     image.save("image.png")
     var = pytesseract.image_to_string(PIL.Image.open('image.png'))
@@ -72,16 +85,7 @@ def ocrSave(sysTrayIcon):
     print("\n\n")
     
 def ocrSearch(sysTrayIcon):
-    clearClip()
-    time.sleep(0.2)
-    image = PIL.ImageGrab.grabclipboard()
-    runSS(None)
-    print("Loading: ", end="")
-    while image is None:
-        print(".", end="")
-        image = PIL.ImageGrab.grabclipboard()
-        time.sleep(0.1)
-    print("\nDone")
+    image = grabSS()
     print("\nText:")
     image.save("image.png")
     var = pytesseract.image_to_string(PIL.Image.open('image.png'))
@@ -90,17 +94,10 @@ def ocrSearch(sysTrayIcon):
     query = var.replace(" ", "+")
     webbrowser.open("http://www.google.co.uk/search?q="+query)
     
+# -------- IMAGE TOOLS
+    
 def imgUpload(sysTrayIcon):
-    clearClip()
-    time.sleep(0.2)
-    image = PIL.ImageGrab.grabclipboard()
-    runSS(None)
-    print("Loading: ", end="")
-    while image is None:
-        print(".", end="")
-        image = PIL.ImageGrab.grabclipboard()
-        time.sleep(0.1)
-    print("\nDone")
+    image = grabSS()
     image.save("image.png")
     headers = {"Authorization": "Client-ID b7cdb801d615f30"}
     j1 = requests.post(
@@ -113,18 +110,8 @@ def imgUpload(sysTrayIcon):
         return False
     webbrowser.open(data["link"])
 
-    
 def imgSearch(sysTrayIcon):
-    clearClip()
-    time.sleep(0.2)
-    image = PIL.ImageGrab.grabclipboard()
-    runSS(None)
-    print("Loading: ", end="")
-    while image is None:
-        print(".", end="")
-        image = PIL.ImageGrab.grabclipboard()
-        time.sleep(0.1)
-    print("\nDone")
+    image = grabSS()
     image.save("image.png")
     filePath = os.path.abspath("image.png")
     multipart = {'encoded_image': (filePath, open(filePath, 'rb')), 'image_content': ''}
@@ -133,33 +120,30 @@ def imgSearch(sysTrayIcon):
     webbrowser.open(fetchUrl)
     
 def imgSave(sysTrayIcon):
-    clearClip()
-    time.sleep(0.2)
-    image = PIL.ImageGrab.grabclipboard()
-    runSS(None)
-    print("Loading: ", end="")
-    while image is None:
-        print(".", end="")
-        image = PIL.ImageGrab.grabclipboard()
-        time.sleep(0.1)
-    print("\nDone")
+    image = grabSS()
     image.save("image.png")
     files = [('PNG Image', '*.png')] 
     file = filedialog.asksaveasfile(filetypes = files, defaultextension = files) 
     image.save(file.name)
-    
-    
+        
+# ------------- SETUP -------------
+        
 menu_options = (
-    ('Screenshot', None, runSS),
-    ('--------------', None, hello),
+    ('Screenshot', None, getSS),
+    ('--------------', None, blank),
     ('Text Copy', None, ocrSave),
     ('Text Search', None, ocrSearch),
-    ('--------------', None, hello),
-    ('Imgur Uplaod', None, imgUpload),
+    ('--------------', None, blank),
+    ('Imgur Upload', None, imgUpload),
     ('Image Search', None, imgSearch),
     ('Image Save', None, imgSave),
-    ('--------------', None, hello))
+    ('--------------', None, blank))
 
 print("\n"*60)
 clearClip()
+
+KeyBinder.activate({
+    'Shift-R': ocrSearch
+}, run_thread=True)
+
 gui.SysTrayIcon(next(icons), hover_text, menu_options, on_quit=close, default_menu_index=1)
